Completed Text Analysis
Text analysis completed, including adding two new functions to deal with removing unneccessary punctuation from char count and creating a final string to be printed out to console. Next step is to do a basic validation of the file to make sure it is a .txt and finish user interaction.